+++ src/commands/filterCoverage.ts
@@ -1,0 +1,9 @@
+import { getContextManager } from '../common';
+
+export function showUnderCoveredClassesCommandHandler(): void {
+  getContextManager().codeCoverageData.setUnderCoveredOnly(true);
+}
+
+export function showAllCoverageCommandHandler(): void {
+  getContextManager().codeCoverageData.setUnderCoveredOnly(false);
+}